requirements.txt
google-genai
python-dotenv
faiss-cpu
sentence-transformers
numpy
pypdf
torch
